requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
DateTime==4.8
docopt==0.6.2
h3==3.7.6
idna==3.4
numpy==1.23.5
peewee==3.15.4
psycopg2-binary==2.9.5
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==13.15
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
timezonefinder==6.1.9
tornado==6.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
yarg==0.1.9
zope.interface==5.5.2
